Introduce `context` package
