+++ api-tests/src/test/kotlin/io/spine/protodata/MemberSpec.kt
@@ -29,6 +29,9 @@ package io.spine.protodata
 import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.string.shouldEndWith
 import io.spine.protodata.backend.ImplicitPluginWithRenderers
+import io.spine.protodata.context.findEnumTypes
+import io.spine.protodata.context.findMessageTypes
+import io.spine.protodata.context.findServices
 import io.spine.protodata.render.Renderer
 import io.spine.protodata.render.SourceFileSet
 import io.spine.protodata.testing.PipelineSetup
