+++ api/src/main/kotlin/io/spine/protodata/context/ContextAware.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright 2023, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -24,34 +24,20 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata
-
-import com.google.common.annotations.VisibleForTesting
-import io.spine.protodata.type.TypeSystem
-import io.spine.server.Closeable
-import io.spine.server.entity.Entity
-import io.spine.server.integration.ThirdPartyContext
-import io.spine.server.query.Querying
+package io.spine.protodata.context
 
 /**
- * A context of code generation.
+ * An object which can be registered with a [CodegenContext].
  */
-public interface CodegenContext : Querying, Closeable {
+public interface ContextAware {
 
     /**
-     * The type system containing all the types available for code generation.
+     * Registers this object with the given [context].
      */
-    public val typeSystem: TypeSystem
+    public fun registerWith(context: CodegenContext)
 
     /**
-     * The `Insertion Points` context which generates events when
-     * [InsertionPoint][io.spine.protodata.renderer.InsertionPoint]s are added to the code.
+     * Tells if this object is registered with a [CodegenContext].
      */
-    public val insertionPointsContext: ThirdPartyContext
-
-    /**
-     * A test-only method which checks if the context has entities of the given type.
-     */
-    @VisibleForTesting
-    public fun <E : Entity<*, *>> hasEntitiesOfType(cls: Class<E>): Boolean
+    public fun isRegistered(): Boolean
 }
