+++ api/src/main/kotlin/io/spine/protodata/plugin/Plugin.kt
@@ -27,8 +27,8 @@
 package io.spine.protodata.plugin
 
 import io.spine.annotation.Internal
-import io.spine.protodata.CodegenContext
 import io.spine.protodata.ConfigurationError
+import io.spine.protodata.context.CodegenContext
 import io.spine.protodata.render.Renderer
 import io.spine.protodata.render.SourceFileSet
 import io.spine.server.BoundedContextBuilder
