+++ api/src/test/kotlin/io/spine/protodata/settings/given/ActionFactorySpecEnv.kt
@@ -30,8 +30,8 @@ import com.google.protobuf.Empty
 import com.google.protobuf.Message
 import com.google.protobuf.StringValue
 import io.spine.base.EntityState
-import io.spine.protodata.CodegenContext
 import io.spine.protodata.MessageType
+import io.spine.protodata.context.CodegenContext
 import io.spine.protodata.render.ActionFactory
 import io.spine.protodata.render.RenderAction
 import io.spine.protodata.render.SourceFile
