+++ backend/src/main/kotlin/io/spine/protodata/backend/Contexts.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -30,8 +30,8 @@ import com.google.common.annotations.VisibleForTesting
 import io.spine.base.EntityState
 import io.spine.base.EventMessage
 import io.spine.core.userId
-import io.spine.protodata.CodegenContext
 import io.spine.protodata.InsertionPointsView
+import io.spine.protodata.context.CodegenContext
 import io.spine.protodata.plugin.add
 import io.spine.protodata.type.TypeSystem
 import io.spine.server.BoundedContext
