+++ backend/src/test/java/io/spine/protodata/type/FakeQuerying.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2023, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -27,7 +27,7 @@
 package io.spine.protodata.type;
 
 import io.spine.base.EntityState;
-import io.spine.protodata.CodegenContext;
+import io.spine.protodata.context.CodegenContext;
 import io.spine.server.query.Querying;
 import io.spine.server.query.QueryingClient;
 import org.checkerframework.checker.nullness.qual.NonNull;
