+++ backend/src/test/kotlin/io/spine/protodata/backend/CodeGenerationContextSpec.kt
@@ -45,18 +45,18 @@ import io.kotest.matchers.string.shouldContain
 import io.spine.option.OptionsProto
 import io.spine.option.OptionsProto.BETA_TYPE_FIELD_NUMBER
 import io.spine.protobuf.AnyPacker
-import io.spine.protodata.CodegenContext
 import io.spine.protodata.PrimitiveType.TYPE_BOOL
 import io.spine.protodata.ProtobufDependency
 import io.spine.protodata.ProtobufSourceFile
 import io.spine.protodata.asType
 import io.spine.protodata.backend.event.CompilerEvents
-import io.spine.protodata.protobuf.file
+import io.spine.protodata.context.CodegenContext
 import io.spine.protodata.option
+import io.spine.protodata.protobuf.file
+import io.spine.protodata.protobuf.toFile
 import io.spine.protodata.test.DoctorProto
 import io.spine.protodata.test.PhDProto
 import io.spine.protodata.test.XtraOptsProto
-import io.spine.protodata.protobuf.toFile
 import io.spine.testing.server.blackbox.BlackBox
 import io.spine.testing.server.blackbox.assertEntity
 import io.spine.time.TimeProto
