+++ backend/src/test/kotlin/io/spine/protodata/backend/PipelineSpec.kt
@@ -33,8 +33,8 @@ import com.google.protobuf.compiler.PluginProtos.CodeGeneratorRequest
 import com.google.protobuf.compiler.codeGeneratorRequest
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldContain
-import io.spine.protodata.CodegenContext
 import io.spine.protodata.ConfigurationError
+import io.spine.protodata.context.CodegenContext
 import io.spine.protodata.render.SourceFileSet
 import io.spine.protodata.render.codeLine
 import io.spine.protodata.settings.Format
