+++ java/src/main/kotlin/io/spine/protodata/java/render/CreateNestedClass.kt
@@ -31,8 +31,8 @@ import com.google.protobuf.stringValue
 import com.intellij.psi.PsiAnnotation
 import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiMethod
-import io.spine.protodata.CodegenContext
 import io.spine.protodata.MessageType
+import io.spine.protodata.context.CodegenContext
 import io.spine.protodata.render.SourceFile
 import io.spine.tools.code.Java
 import io.spine.tools.psi.java.Environment.elementFactory
