+++ java/src/main/kotlin/io/spine/protodata/java/render/ImplementInterface.kt
@@ -26,8 +26,8 @@
 
 package io.spine.protodata.java.render
 
-import io.spine.protodata.CodegenContext
 import io.spine.protodata.MessageType
+import io.spine.protodata.context.CodegenContext
 import io.spine.protodata.java.ClassName
 import io.spine.protodata.render.SourceFile
 import io.spine.tools.code.Java
